// src/types.d.ts
declare enum NOTIFICATION_CONTAINER {
    BOTTOM_LEFT = "bottom-left",
    BOTTOM_RIGHT = "bottom-right",
    BOTTOM_CENTER = "bottom-center",
    TOP_LEFT = "top-left",
    TOP_RIGHT = "top-right",
    TOP_CENTER = "top-center",
    CENTER = "center",
    TOP_FULL = "top-full",
    BOTTOM_FULL = "bottom-full"
}
declare enum NOTIFICATION_INSERTION {
    TOP = "top",
    BOTTOM = "bottom"
}
declare enum NOTIFICATION_TYPE {
    SUCCESS = "success",
    DANGER = "danger",
    INFO = "info",
    DEFAULT = "default",
    WARNING = "warning"
}
// ...
interface iNotification {
    id?: string;
    onRemoval?: (id: string, removalFlag: string) => void;
    title?: NotificationTitleMessage;
    message?: NotificationTitleMessage;
    content?: NotificationContent;
    type?: NOTIFICATION_TYPE;
    container?: NOTIFICATION_CONTAINER;
    insert?: NOTIFICATION_INSERTION;
    dismiss?: iNotificationDismiss;
    animationIn?: string[];
    animationOut?: string[];
    slidingEnter?: iTransition;
    slidingExit?: iTransition;
    touchRevert?: iTransition;
    touchSlidingExit?: {
        fade: iTransition;
        swipe: iTransition;
    };
    userDefinedTypes?: iNotificationCustomType[];
    width?: number;
    hasBeenRemoved?: boolean;
}
declare type NotificationTitleMessage = string | React.ReactNode | React.FunctionComponent;
declare type NotificationContent = React.ComponentClass<any, any> | React.FunctionComponent<any> | React.ReactElement;
interface iTransition {
    duration: number;
    timingFunction: string;
    delay: number;
}
interface iTouchTransition {
    swipe: iTransition;
    fade: iTransition;
}
interface iNotificationDismiss {
    duration: number;
    onScreen: boolean;
    pauseOnHover: boolean;
    waitForAnimation: boolean;
    click: boolean;
    touch: boolean;
    showIcon: boolean;
}
interface iNotificationCustomType {
    htmlClasses: string[];
    name: string;
}